+++ src/towerdefense/game/projectiles/Glue.java
@@ -11,7 +11,7 @@ public class Glue extends Projectile {
     private static final String graphicsPath = "towers/glue_tower/projectile.png";
     private static final double proportion = 1.0 / 5.0;
 
-    public Glue(Map map, GameModel gameModel, int damage, int velocity, Position initialPosition, NPC target) {
+    public Glue(Map map, GameModel gameModel, int damage, double velocity, Position initialPosition, NPC target) {
         super(map, gameModel, damage, velocity, initialPosition, target);
     }
